Azure custome config resolver
