+++ src/AzureCloud/AzureResourceOptions.cs
@@ -5,18 +5,36 @@ using Microsoft.Extensions.Configuration;
 
 namespace Squadron.AzureCloud
 {
+    public class AzureResourceOptionsBuilder
+    {
+        internal Func<AzureResourceConfiguration> ConfigResolver { get;  set; }
+
+        public AzureResourceOptionsBuilder()
+        {
+            ConfigResolver = AzureResourceOptions.DefaultAzureConfigurationResolver;
+        }
+
+        public AzureResourceOptionsBuilder SetConfigResolver(
+            Func<AzureResourceConfiguration> resolver)
+        {
+            ConfigResolver = resolver;
+            return this;
+        }
+    }
+
+
     /// <summary>
     /// Base options to use with Azure resources
     /// </summary>
-    /// <seealso cref="Squadron.AzureCloud.IAzureResourceConfigurationProvider" />
-    public class AzureResourceOptions : IAzureResourceConfigurationProvider
+    public class AzureResourceOptions 
     {
+        public Func<AzureResourceConfiguration> ConfigResolver { get; set; }
 
         /// <summary>
         /// Gets the azure configuration using the .NET configuration system
         /// </summary>
         /// <returns></returns>
-        public AzureResourceConfiguration GetAzureConfiguration()
+        internal static AzureResourceConfiguration DefaultAzureConfigurationResolver()
         {
             IConfigurationRoot configuration = new ConfigurationBuilder()
                 .AddJsonFile("appsettings.json", true)
@@ -24,8 +42,8 @@ namespace Squadron.AzureCloud
                 .AddEnvironmentVariables()
                 .Build();
 
-            IConfigurationSection section = configuration.GetSection("Squadron:Azure");
 
+            IConfigurationSection section = configuration.GetSection("Squadron:Azure");
             AzureResourceConfiguration azureConfig = section.Get<AzureResourceConfiguration>();
             return azureConfig;
         }
