+++ src/AzureCloudServiceBus/AzureCloudServiceBusOptions.cs
@@ -17,5 +17,6 @@ namespace Squadron
         /// </summary>
         /// <param name="builder">The builder.</param>
         public abstract void Configure(ServiceBusOptionsBuilder builder);
+
     }
 }
